test: split integration specs across files
Split the largest Playwright integration specs into smaller behavior-focused files.

This lets the existing file-level parallelism schedule more of the suite concurrently.

Add local helpers for shared statistics, spoiler fixture, and boot-reconcile setup.

Move keyboard shortcut coverage into focused stats and reader specs after the split.
