+++ tests/integration/reader/image-long-press.spec.ts
@@ -1,11 +1,6 @@
 import type { Locator, Page } from '@playwright/test';
 import { expect, test } from '../helpers/harness.ts';
-import {
-  expectBookReaderText,
-  importBookFixtures,
-  openBookFromManage,
-  SPOILER_IMAGE_GALLERY_BOOK
-} from '../helpers/fixtures.ts';
+import { openBookFromManage, SPOILER_IMAGE_GALLERY_BOOK } from '../helpers/fixtures.ts';
 import {
   readerImage,
   readerImageURL,
@@ -13,59 +8,11 @@ import {
   readerSVGImageURL
 } from '../helpers/reader.ts';
 import { useReaderSettings } from '../helpers/workflows.ts';
+import { openSpoilerFixtureBook } from './spoiler-fixture.ts';
 
 const COVER_ALT = 'Cover before spoilers';
 const SVG_LABEL = 'SVG before spoilers';
 const SPOILER_ONE_ALT = 'Spoiler illustration one';
-
-test('book content enhancement wires generated anchors and ruby clicks', async ({ page }) => {
-  await openSpoilerFixtureBook(page, {
-    furigana: 'Toggle',
-    writingMode: 'Horizontal'
-  });
-
-  await page.getByRole('link', { name: 'Ruby sample' }).click();
-
-  const rubySample = page.locator('#ruby-sample');
-  await expect(rubySample).toBeInViewport();
-
-  const ruby = rubySample.locator('ruby');
-  const rubyText = ruby.locator('rt');
-  await expect(rubyText).toBeHidden();
-  await ruby.dispatchEvent('click');
-  await expect(rubyText).toBeVisible();
-  await ruby.dispatchEvent('click');
-  await expect(rubyText).toBeHidden();
-});
-
-test('furigana setting changes update ruby click behavior after reopening the reader', async ({
-  page
-}) => {
-  await useReaderSettings(page, {
-    furigana: 'Default',
-    viewMode: 'Continuous',
-    writingMode: 'Horizontal'
-  });
-  await importBookFixtures(page, [SPOILER_IMAGE_GALLERY_BOOK]);
-  await openBookFromManage(page, SPOILER_IMAGE_GALLERY_BOOK);
-  await openRubySample(page);
-
-  const defaultRuby = page.locator('#ruby-sample ruby');
-  const defaultRubyText = defaultRuby.locator('rt');
-  await expect(defaultRubyText).toBeVisible();
-  await defaultRuby.dispatchEvent('click');
-  await expect(defaultRubyText).toBeVisible();
-
-  await useReaderSettings(page, { furigana: 'Toggle' });
-  await openBookFromManage(page, SPOILER_IMAGE_GALLERY_BOOK);
-  await openRubySample(page);
-
-  const toggleRuby = page.locator('#ruby-sample ruby');
-  const toggleRubyText = toggleRuby.locator('rt');
-  await expect(toggleRubyText).toBeHidden();
-  await toggleRuby.dispatchEvent('click');
-  await expect(toggleRubyText).toBeVisible();
-});
 
 test('normal images do not get long-press open behavior outside PWA display mode', async ({
   page
@@ -135,24 +82,6 @@ test('spoiler image long-press follows the blur image setting and reveal state',
     await readerImageURL(page, SPOILER_ONE_ALT)
   );
 });
-
-async function openSpoilerFixtureBook(
-  page: Page,
-  settings: Parameters<typeof useReaderSettings>[1] = {}
-) {
-  await useReaderSettings(page, {
-    viewMode: 'Continuous',
-    ...settings
-  });
-  await importBookFixtures(page, [SPOILER_IMAGE_GALLERY_BOOK]);
-  await openBookFromManage(page, SPOILER_IMAGE_GALLERY_BOOK);
-  await expectBookReaderText(page, SPOILER_IMAGE_GALLERY_BOOK);
-}
-
-async function openRubySample(page: Page) {
-  await page.getByRole('link', { name: 'Ruby sample' }).click();
-  await expect(page.locator('#ruby-sample')).toBeInViewport();
-}
 
 async function expectLongPressToOpen(page: Page, locator: Locator, expectedURL: string) {
   const center = await moveMouseToCenter(page, locator);
