--- tests/integration/statistics/actions.spec.ts
@@ -1,322 +1,0 @@
-import { stat } from 'node:fs/promises';
-import type { Locator, Page } from '@playwright/test';
-import { expect, SYNC_ASSERTION_TIMEOUT, test } from '../helpers/harness.ts';
-import {
-  fixtureTitle,
-  importBookFixtures,
-  LONG_BOOK,
-  openBookFromManage,
-  openStatisticsSettings,
-  PLAIN_TEXT_BOOK,
-  recordStatisticForBook,
-  VALID_BOOK,
-  type LibraryBookFixture
-} from '../helpers/fixtures.ts';
-import { navigateToStatisticsSummary } from '../helpers/navigation.ts';
-import { showReaderHeader } from '../helpers/reader.ts';
-import { enableStatistics } from '../helpers/workflows.ts';
-
-const EARLIER_STAT_DATE = '2026-04-10';
-const LATER_STAT_DATE = '2026-05-10';
-
-test('statistics header and settings actions operate on loaded statistics', async ({
-  context,
-  page
-}, testInfo) => {
-  const exportPath = testInfo.outputPath('statistics-export.zip');
-
-  await context.grantPermissions(['clipboard-read', 'clipboard-write']);
-  await page.clock.install({ time: new Date(`${EARLIER_STAT_DATE}T11:59:00Z`) });
-  await enableStatistics(page);
-  await importBookFixtures(page, [LONG_BOOK, VALID_BOOK]);
-  await recordStatisticForBook(page, LONG_BOOK, EARLIER_STAT_DATE);
-  await recordStatisticForBook(page, VALID_BOOK, EARLIER_STAT_DATE, { durationMs: 61_000 });
-  await recordStatisticForBook(page, VALID_BOOK, LATER_STAT_DATE, { durationMs: 61_000 });
-
-  await navigateToStatisticsSummary(page);
-  await expect(page.getByText(LATER_STAT_DATE, { exact: true })).toBeVisible({
-    timeout: SYNC_ASSERTION_TIMEOUT
-  });
-  await expect(page.getByText(EARLIER_STAT_DATE, { exact: true })).toHaveCount(0);
-
-  await page.getByRole('button', { name: 'Filter', exact: true }).click();
-  const filter = page.locator('dialog.sidebar-overlay[open]').filter({
-    has: page.getByPlaceholder('Filter book list')
-  });
-  await expect(filter).toBeVisible();
-  await expect(filter.getByText(fixtureTitle(LONG_BOOK), { exact: true })).toBeVisible();
-  await filter
-    .locator('label')
-    .filter({ hasText: 'Only show books with statistics in the target date range' })
-    .click();
-  await expect(filter.getByText(fixtureTitle(LONG_BOOK), { exact: true })).toHaveCount(0);
-  await filter.getByTitle('Close book filter').click();
-  await expect(filter).toHaveCount(0, { timeout: SYNC_ASSERTION_TIMEOUT });
-
-  const settings = await openStatisticsSettings(page);
-  await settings.getByLabel('Template').selectOption('This Year');
-  await expect(page.getByText(EARLIER_STAT_DATE, { exact: true })).toBeVisible({
-    timeout: SYNC_ASSERTION_TIMEOUT
-  });
-
-  await settings.getByRole('button', { name: 'Set to all time for the selected books' }).click();
-  await expect(page.getByText(EARLIER_STAT_DATE, { exact: true })).toBeVisible({
-    timeout: SYNC_ASSERTION_TIMEOUT
-  });
-  await settings.getByTitle('Close statistics settings').click();
-  await expect(settings).toHaveCount(0, { timeout: SYNC_ASSERTION_TIMEOUT });
-
-  await page.getByRole('button', { name: /Copy/ }).click();
-  await page.getByRole('button', { name: 'Reading Time' }).click();
-  await expect
-    .poll(() => page.evaluate(() => navigator.clipboard.readText()), {
-      timeout: SYNC_ASSERTION_TIMEOUT
-    })
-    .toContain(`.log readtime 2 ${fixtureTitle(VALID_BOOK)}`);
-
-  const exportSettings = await openStatisticsSettings(page);
-  const [download] = await Promise.all([
-    page.waitForEvent('download'),
-    exportSettings.getByRole('button', { name: 'Export Selection' }).click()
-  ]);
-  expect(download.suggestedFilename()).toMatch(/^miwake-reader-export-[\d-]+\.zip$/);
-  await download.saveAs(exportPath);
-  expect((await stat(exportPath)).size).toBeGreaterThan(0);
-
-  await exportSettings.getByRole('button', { name: 'Delete All' }).click();
-  const dialog = page.locator('dialog[open]').filter({
-    has: page.getByRole('heading', { name: 'Delete data' })
-  });
-  await dialog.getByRole('button', { name: 'Delete', exact: true }).click();
-  await expect(page.getByText(/No Data found/)).toBeVisible({ timeout: SYNC_ASSERTION_TIMEOUT });
-});
-
-test('reader statistics navigation preserves the active tab and filters to the current book', async ({
-  page
-}) => {
-  await page.clock.install({ time: new Date(`${LATER_STAT_DATE}T11:59:00Z`) });
-  await enableStatistics(page);
-  await importBookFixtures(page, [LONG_BOOK, PLAIN_TEXT_BOOK, VALID_BOOK]);
-  await recordStatisticForBook(page, LONG_BOOK, LATER_STAT_DATE);
-  await recordStatisticForBook(page, PLAIN_TEXT_BOOK, LATER_STAT_DATE);
-  await recordStatisticForBook(page, VALID_BOOK, LATER_STAT_DATE);
-
-  await navigateToStatisticsSummary(page);
-  await page.getByRole('button', { name: 'Heatmap', exact: true }).click();
-  await expectStatisticsView(page, 'heatmap');
-
-  await openBookFromManage(page, PLAIN_TEXT_BOOK);
-  const header = await showReaderHeader(page);
-  await header.getByRole('button', { name: 'Statistics', exact: true }).click();
-  await expect(page).toHaveURL(/\/statistics/);
-  await expectStatisticsView(page, 'heatmap');
-  await expect
-    .poll(() => new URL(page.url()).searchParams.get('b'), {
-      timeout: SYNC_ASSERTION_TIMEOUT
-    })
-    .toMatch(/^\d+$/);
-
-  await page.getByRole('button', { name: 'Summary', exact: true }).click();
-  await expectStatisticsView(page, 'summary');
-
-  await expectSummaryBookVisible(page, PLAIN_TEXT_BOOK);
-  await expectSummaryBookHidden(page, LONG_BOOK);
-  await expectSummaryBookHidden(page, VALID_BOOK);
-
-  const filter = await openStatisticsFilter(page);
-  await setStatisticsFilterBook(filter, LONG_BOOK, true);
-  await expectStatisticsBookFilterCount(page, 2);
-  await filter.getByTitle('Close book filter').click();
-  await expect(filter).toHaveCount(0, { timeout: SYNC_ASSERTION_TIMEOUT });
-
-  await expectSummaryBookVisible(page, PLAIN_TEXT_BOOK);
-  await expectSummaryBookVisible(page, LONG_BOOK);
-  await expectSummaryBookHidden(page, VALID_BOOK);
-
-  await page.reload();
-  await expectStatisticsBookFilterCount(page, 2);
-
-  await expectSummaryBookVisible(page, PLAIN_TEXT_BOOK);
-  await expectSummaryBookVisible(page, LONG_BOOK);
-  await expectSummaryBookHidden(page, VALID_BOOK);
-});
-
-test('statistics keyboard shortcuts change the range template and aggregation mode', async ({
-  page
-}) => {
-  await page.clock.install({ time: new Date(`${LATER_STAT_DATE}T11:59:00Z`) });
-  await enableStatistics(page);
-  await importBookFixtures(page, [VALID_BOOK]);
-  await recordStatisticForBook(page, VALID_BOOK, LATER_STAT_DATE, { durationMs: 61_000 });
-
-  await navigateToStatisticsSummary(page);
-
-  const summary = page.getByRole('region', { name: 'Statistics summary' });
-  await expect(page.getByText(`Data for ${LATER_STAT_DATE}`, { exact: true })).toBeVisible({
-    timeout: SYNC_ASSERTION_TIMEOUT
-  });
-  await expect(summary.getByText('Book', { exact: true }).first()).toBeVisible();
-
-  await page.keyboard.press('Shift+T');
-  await expect(page.getByText(`Data for ${LATER_STAT_DATE}`, { exact: true })).toBeVisible();
-
-  await page.keyboard.down('t');
-  await expect(page.getByText('Data for 2026-05-04 - 2026-05-10', { exact: true })).toBeVisible({
-    timeout: SYNC_ASSERTION_TIMEOUT
-  });
-  await page.keyboard.down('t');
-  await page.keyboard.up('t');
-  await expect(page.getByText('Data for 2026-05-04 - 2026-05-10', { exact: true })).toBeVisible();
-
-  await page.keyboard.press('Shift+A');
-  await expect(summary.getByText('Book', { exact: true }).first()).toBeVisible();
-
-  await page.keyboard.down('a');
-  await expect(summary.getByText('Book', { exact: true }).first()).toBeHidden();
-  await page.keyboard.down('a');
-  await page.keyboard.up('a');
-  await expect(summary.getByText('Book', { exact: true }).first()).toBeHidden();
-
-  const settings = await openStatisticsSettings(page);
-  await expect(settings.getByLabel('Template')).toHaveValue('This Week');
-  await expect(settings.getByLabel('Primary Aggregration')).toHaveValue('Date');
-});
-
-test('statistics summary headers stay sticky while the page scrolls', async ({ page }) => {
-  await page.setViewportSize({ width: 900, height: 360 });
-  await page.clock.install({ time: new Date(`${LATER_STAT_DATE}T11:59:00Z`) });
-  await enableStatistics(page);
-  await importBookFixtures(page, [VALID_BOOK]);
-
-  for (let day = 1; day <= 6; day += 1) {
-    await recordStatisticForBook(page, VALID_BOOK, `2026-05-${day.toString().padStart(2, '0')}`);
-  }
-
-  await navigateToStatisticsSummary(page);
-  const settings = await openStatisticsSettings(page);
-  await settings.getByRole('button', { name: 'Set to all time for the selected books' }).click();
-  await settings.getByTitle('Close statistics settings').click();
-  await expect(settings).toHaveCount(0, { timeout: SYNC_ASSERTION_TIMEOUT });
-
-  const summary = page.getByRole('region', { name: 'Statistics summary' });
-  await expect
-    .poll(() => page.evaluate(() => document.scrollingElement!.scrollHeight > window.innerHeight), {
-      timeout: SYNC_ASSERTION_TIMEOUT
-    })
-    .toBe(true);
-  await expect
-    .poll(() => summary.evaluate((el) => el.scrollHeight - el.clientHeight), {
-      timeout: SYNC_ASSERTION_TIMEOUT
-    })
-    .toBeLessThanOrEqual(1);
-
-  const readingTimeHeader = summary.getByText('Total Time', { exact: true });
-  const headerTop = await readingTimeHeader.evaluate((el) =>
-    Math.round(el.getBoundingClientRect().top)
-  );
-
-  await page.evaluate(() => window.scrollTo(0, document.scrollingElement!.scrollHeight));
-
-  await expect
-    .poll(() => readingTimeHeader.evaluate((el) => Math.round(el.getBoundingClientRect().top)), {
-      timeout: SYNC_ASSERTION_TIMEOUT
-    })
-    .toBeLessThan(headerTop);
-  await expect
-    .poll(() => readingTimeHeader.evaluate((el) => Math.round(el.getBoundingClientRect().top)), {
-      timeout: SYNC_ASSERTION_TIMEOUT
-    })
-    .toBeGreaterThanOrEqual(48);
-  await expect
-    .poll(() => readingTimeHeader.evaluate((el) => Math.round(el.getBoundingClientRect().top)), {
-      timeout: SYNC_ASSERTION_TIMEOUT
-    })
-    .toBeLessThanOrEqual(72);
-});
-
-test('statistics heatmap expands day cells to fill desktop width', async ({ page }) => {
-  await page.setViewportSize({ width: 1440, height: 900 });
-  await page.clock.install({ time: new Date(`${LATER_STAT_DATE}T11:59:00Z`) });
-  await enableStatistics(page);
-  await importBookFixtures(page, [VALID_BOOK]);
-  await recordStatisticForBook(page, VALID_BOOK, LATER_STAT_DATE);
-
-  await navigateToStatisticsSummary(page);
-  await page.getByRole('button', { name: 'Heatmap', exact: true }).click();
-
-  const heatmap = page.getByRole('grid', { name: 'Reading Data for 2026' });
-  const dayCell = heatmap.getByRole('cell', { name: new RegExp(`^${LATER_STAT_DATE},`) });
-  await expect(dayCell).toBeVisible({ timeout: SYNC_ASSERTION_TIMEOUT });
-
-  const { actualWidth, expectedWidth } = await dayCell.evaluate((el) => {
-    const heatmapGrid = el.closest('[role="grid"]');
-    if (!(heatmapGrid instanceof HTMLElement) || !heatmapGrid.parentElement) {
-      throw new Error('Could not find heatmap container');
-    }
-
-    const containerWidth = heatmapGrid.parentElement.clientWidth;
-    const arrowElementsWidth = 30;
-    const gridGap = 2;
-    const gridColumnsPerYear = 54;
-    const gridColumnsWidth = gridColumnsPerYear * gridGap;
-    const dayGridColumns = 3;
-    const allGridColumns = gridColumnsPerYear + dayGridColumns;
-
-    return {
-      actualWidth: el instanceof HTMLElement ? el.offsetWidth : el.getBoundingClientRect().width,
-      expectedWidth: Math.max(
-        15,
-        (containerWidth - arrowElementsWidth - gridColumnsWidth) / allGridColumns
-      )
-    };
-  });
-
-  expect(actualWidth).toBeGreaterThanOrEqual(expectedWidth - 1);
-});
-
-async function openStatisticsFilter(page: Page) {
-  await page.getByRole('button', { name: 'Filter', exact: true }).click();
-  const filter = page.locator('dialog.sidebar-overlay[open]').filter({
-    has: page.getByPlaceholder('Filter book list')
-  });
-  await expect(filter).toBeVisible({ timeout: SYNC_ASSERTION_TIMEOUT });
-  return filter;
-}
-
-async function setStatisticsFilterBook(
-  filter: Locator,
-  fixture: LibraryBookFixture,
-  checked: boolean
-) {
-  await filter
-    .locator('tr')
-    .filter({ hasText: fixtureTitle(fixture) })
-    .getByRole('checkbox')
-    .setChecked(checked);
-}
-
-async function expectStatisticsBookFilterCount(page: Page, count: number) {
-  await expect
-    .poll(() => new URL(page.url()).searchParams.getAll('b').length, {
-      timeout: SYNC_ASSERTION_TIMEOUT
-    })
-    .toBe(count);
-}
-
-async function expectStatisticsView(page: Page, view: 'summary' | 'heatmap') {
-  await expect
-    .poll(() => new URL(page.url()).searchParams.get('view'), {
-      timeout: SYNC_ASSERTION_TIMEOUT
-    })
-    .toBe(view);
-}
-
-async function expectSummaryBookVisible(page: Page, fixture: LibraryBookFixture) {
-  await expect(page.getByTitle(fixtureTitle(fixture)).filter({ visible: true })).toBeVisible({
-    timeout: SYNC_ASSERTION_TIMEOUT
-  });
-}
-
-async function expectSummaryBookHidden(page: Page, fixture: LibraryBookFixture) {
-  await expect(page.getByTitle(fixtureTitle(fixture)).filter({ visible: true })).toHaveCount(0);
-}
